SRB2: MSVC6 warning cleanup
git-svn-id: svn://code.srb2.org/SRB2/trunk@3799 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/d_clisrv.c
@@ -198,7 +198,7 @@ void SendNetXCmd2(netxcmd_t id, const void *param, size_t nparam)
 byte GetFreeXCmdSize(void)
 {
 	// -1 for the size and another -1 for the ID.
-	return localtextcmd[0] - 2;
+	return (byte)(localtextcmd[0] - 2);
 }
 
 static void ExtraDataTicker(void)
